use same gas amount as in rust

+++ packages/arb-avm-cpp/avm/include/avm/machinestate/machinestate.hpp
@@ -30,7 +30,7 @@ struct AssertionContext {
     std::vector<Tuple> inbox_messages;
     uint64_t inbox_messages_consumed;
     Tuple sideload_value;
-    uint32_t numSteps;
+    uint64_t numSteps;
     uint64_t numGas;
     bool blockingSideload;
     nonstd::optional<value> fake_inbox_peek_value;
